requirements.txt
python-ldap 
MySQL-python